requirements.txt
flask==3.0.3
pycryptodome==3.20.0
requests==2.32.3
